@@ -2,12 +2,9 @@
 
   <Import Project="..\..\tests.props" />
   <PropertyGroup>
-    <TargetFrameworks>net6.0;net8.0</TargetFrameworks>
+    <TargetFrameworks>net8.0</TargetFrameworks>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Routing.Abstractions" Version="2.2.0" />
-  </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\Candoumbe.MiscUtilities\Candoumbe.MiscUtilities.csproj" />
   </ItemGroup>
